requirements.txt
mysql-connector-python
flask
flask_cors
werkzeug
